Added a module 'common_utility_resources' as an Android Module to provide utilitarian methods that helps getting access to the files from resource assets of the Project. Added a module 'common_utility_views' as an Android Module to provide utilitarian methods around Views. And since, these two modules are a result of separating out the individual components, the 'common_utility' Module is affected as well.
