+++ common_utility_resources/.gitignore
@@ -1,0 +1,1 @@
+/build
