+++ settings.gradle
@@ -6,4 +6,6 @@ include ':common_data_ui'
 include ':common_data_network'
 include ':common_utility_permissions'
 include ':common_utility_network_callback'
+include ':common_utility_resources'
+include ':common_utility_views'
 include ':common_utility'
